Inline RequestBodyCommon
This is an artifact of our multiplatform exploration.
